% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_service.R
\name{build_service}
\alias{build_service}
\title{build_service}
\usage{
build_service(dns_names = NULL, stages = NULL)
}
\arguments{
\item{dns_names}{The hostnames of the services to run code on}

\item{stages}{Names for each of the services.}
}
\description{
The workhorse
}
