% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_service.R
\name{rebuild_service}
\alias{rebuild_service}
\title{rebuild_service}
\usage{
rebuild_service(dns_names = NULL, stages = NULL)
}
\arguments{
\item{dns_names}{The hostnames of the services to run code on}

\item{stages}{Names for each of the services.}
}
\description{
Rebuild the service
}
